build warnings

+++ src/switch_event.c
@@ -299,7 +299,7 @@ static switch_status_t switch_event_queue_dispatch_event(switch_event_t **eventp
 
 		switch_mutex_lock(EVENT_QUEUE_MUTEX);		
 
-		if (switch_queue_size(EVENT_DISPATCH_QUEUE) > (DISPATCH_QUEUE_LEN * DISPATCH_THREAD_COUNT)) {
+		if (switch_queue_size(EVENT_DISPATCH_QUEUE) > (unsigned int)(DISPATCH_QUEUE_LEN * DISPATCH_THREAD_COUNT)) {
 			launch++;
 		}
 		
@@ -467,7 +467,7 @@ SWITCH_DECLARE(switch_status_t) switch_event_shutdown(void)
 	switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_CONSOLE, "Stopping dispatch queues\n");
 
 	
-	for(x = 0; x < DISPATCH_THREAD_COUNT; x++) {
+	for(x = 0; x < (uint32_t)DISPATCH_THREAD_COUNT; x++) {
 		switch_queue_trypush(EVENT_DISPATCH_QUEUE, NULL);
 	}
 
@@ -475,7 +475,7 @@ SWITCH_DECLARE(switch_status_t) switch_event_shutdown(void)
 
 	switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_CONSOLE, "Stopping dispatch threads\n");
 
-	for(x = 0; x < DISPATCH_THREAD_COUNT; x++) {
+	for(x = 0; x < (uint32_t)DISPATCH_THREAD_COUNT; x++) {
 		switch_status_t st;
 		switch_thread_join(&st, EVENT_DISPATCH_QUEUE_THREADS[x]);
 	}
